requirements.txt
beautifulsoup4
bs4
certifi
charset-normalizer
idna
requests
soupsieve
urllib3
python-dotenv
